Move jQuery.easing functionality into integration module
