+++ js/integration/jquery.js
@@ -5,4 +5,5 @@ var jquery = require("jquery");
 // Check availability in global environment
 if(jquery) {
     require("./jquery/renderer");
+    require("./jquery/easing");
 }
