+++ js/integration/jquery/easing.js
@@ -1,0 +1,6 @@
+"use strict";
+
+var jquery = require("jquery"),
+    easing = require("../../animation/easing");
+
+easing.setEasing(jquery.easing);
