+++ testing/tests/DevExpress.animation/fx.tests.js
@@ -2,7 +2,6 @@
 
 var $ = require("jquery"),
     fx = require("animation/fx"),
-    internals = fx.__internals,
     translator = require("animation/translator"),
     animationFrame = require("animation/frame"),
     support = require("core/utils/support"),
@@ -1031,28 +1030,6 @@ if(support.transition) {
         fx.stop($element);
     });
 }
-
-
-
-QUnit.module("easing");
-
-QUnit.test("css transition timing func parsing", function(assert) {
-    var convert = internals.convertTransitionTimingFuncToJQueryEasing;
-
-    var easingName = convert("linear");
-    assert.equal(easingName, "cubicbezier_0_0_1_1");
-
-    easingName = convert("ease-in-out");
-    assert.equal(easingName, "cubicbezier_0p42_0_0p58_1");
-
-    easingName = convert("cubic-bezier(0.190, 1.000, 0.220, 1.000)");
-    assert.equal(easingName, "cubicbezier_0p19_1_0p22_1");
-
-    assert.equal($.easing["cubicbezier_0p19_1_0p22_1"](0.0056, 14, 0, 1, 2500).toFixed(3), 0.029);
-    assert.equal($.easing["cubicbezier_0p19_1_0p22_1"](0.156, 390, 0, 1, 2500).toFixed(3), 0.667);
-    assert.equal($.easing["cubicbezier_0p19_1_0p22_1"](0.344, 860, 0, 1, 2500).toFixed(3), 0.924);
-    assert.equal($.easing["cubicbezier_0p19_1_0p22_1"](0.9864, 2466, 0, 1, 2500).toFixed(3), 1);
-});
 
 
 QUnit.module("fx effects");
